Added transcriber, video/audio features, OCR, and updated functions

+++ updated demo wip/app.py
@@ -1,9 +1,9 @@
 from flask import *
-import os
-import tempfile
 import whisper
 import pymupdf4llm as p4
 import nltk
+import tempfile
+import os
 from transformers import pipeline
 from pypdf import PdfReader
 from google import genai
@@ -11,7 +11,15 @@ import base64
 
 app = Flask(__name__)
 
-client = genai.Client(api_key="INSERT-YOUR-KEY")
+client = genai.Client(api_key="AIzaSyBPj27ASsz9FmiTt0vn3sy-HsMYCWlpUqE")
+model = whisper.load_model("base")
+
+def clean(output):
+    output = output.replace('##', '')
+    output = output.replace('**', '')
+    output = output.replace('\n', '<br>')
+
+    return output
 
 def summarize(text):
     summary_response = client.models.generate_content(
@@ -35,65 +43,62 @@ def generate_quiz(text):
     )
     return quiz_response.text
 
+def extract_text(encoded_img):
+    ocr_response = client.models.generate_content(
+        model="gemini-2.0-flash",
+        contents=[{
+            "parts": [
+                {"text": "Tell me what this image says and nothing else"},
+                {"inline_data": {
+                    "mime_type": "image/png", #change to image/png,webp,etc. 
+                    "data": encoded_img
+                }}
+            ]
+        }]
+    )
+    return ocr_response.text
+
 def get_doc_summary(name):
-    reader = p4.LlamaMarkdownReader()
-    data = reader.load_data(os.path.abspath(name))
 
+    reader = PdfReader(name)
     text = ""
-    for i in range(len(data)):
-        text += data[i].text_resource.text
-    lines = text.split("\n")
-    for i in range(len(lines)):
-        lines[i] = lines[i].replace("##", "")
-        lines[i] = lines[i].replace("**", "")
-        lines[i] = lines[i].replace("\u200b", "-->")
-        if (lines[i] != "" and lines[i][0] == "#"):
-            lines[i] = lines[i][1::]
-        lines[i] = lines[i].strip().rstrip()
+    for page in reader.pages:
+        text += page.extract_text()
+
+    summary = summarize(text)
+    return clean(summary)
+
+def get_transcript(name):
+    return model.transcribe(name)['text']
+
+def get_vid_aud_summary(name):
+    transcript = get_transcript(name)
     
-    ret = []
-    for i in range(len(lines)):
-        if (lines[i].lower() in ["references", "works cited"]):
-            break
-        if (lines[i] != ''):
-            ret.append(lines[i])
-    lines = ret
+    summary = summarize(transcript)
+    return clean(summary)
 
-    t = ""
-    for line in lines:
-        t += line + "\n"
+def get_vid_aud_quiz(name):
+    transcript = get_transcript(name)
 
-    return summarize(t)
+    quiz = generate_quiz(transcript)
+    return clean(quiz)
 
 def get_doc_quiz(name):
-    reader = p4.LlamaMarkdownReader()
-    data = reader.load_data(os.path.abspath(name))
 
+    reader = PdfReader(name)
     text = ""
-    for i in range(len(data)):
-        text += data[i].text_resource.text
-    lines = text.split("\n")
-    for i in range(len(lines)):
-        lines[i] = lines[i].replace("##", "")
-        lines[i] = lines[i].replace("**", "")
-        lines[i] = lines[i].replace("\u200b", "-->")
-        if (lines[i] != "" and lines[i][0] == "#"):
-            lines[i] = lines[i][1::]
-        lines[i] = lines[i].strip().rstrip()
-    
-    ret = []
-    for i in range(len(lines)):
-        if (lines[i].lower() in ["references", "works cited"]):
-            break
-        if (lines[i] != ''):
-            ret.append(lines[i])
-    lines = ret
+    for page in reader.pages:
+        text += page.extract_text()
 
-    t = ""
-    for line in lines:
-        t += line + "\n"
+    quiz = generate_quiz(text)
+    return clean(quiz)
 
-    return generate_quiz(t)
+def get_img_txt(name):
+
+    encoded_image_string = base64.b64encode(name.read()).decode("utf-8")
+
+    text = extract_text(encoded_image_string)
+    return clean(text)
 
 @app.route("/")
 def home():
@@ -103,54 +108,86 @@ def home():
 def summary():
     return render_template('summary.html')
 
-@app.route("/summary-doc", methods=["POST", "GET"])
-def summary_doc():
+@app.route("/summary-result", methods=["POST", "GET"])
+def summary_result():
     if (request.method == "POST"):
-
         file_upload = request.files.get("file")
-        
-        with tempfile.NamedTemporaryFile(delete=False, suffix=os.path.splitext(file_upload.filename)[1]) as tmp:
-            file_upload.save(tmp.name)
-            tmp_path = tmp.name
-        
-        try:
-            result = get_doc_summary(tmp_path).replace('\n', '<br>')
 
-            template = f"<p class='animated' id='summary'>{result}</p>"
-
-            return render_template('summary-doc.html') + template
+        if file_upload.mimetype.startswith("audio/") or file_upload.mimetype.startswith("video/"):
+            with tempfile.NamedTemporaryFile(delete=False, suffix=os.path.splitext(file_upload.filename)[1]) as tmp:
+                file_upload.save(tmp.name)
+                tmp_path = tmp.name
+            try:
+                result = get_vid_aud_summary(tmp_path)
+            finally:
+                os.remove(tmp_path)
+        elif file_upload.mimetype == "application/pdf":
+            result = get_doc_summary(file_upload)
+            
+        template = f"<p class='animated' id='summary'>{result}</p>"
+        return render_template('summary-result.html') + template
     
-        finally:
-            os.remove(tmp_path)
-    
-    return render_template("summary-doc.html")
+    return render_template("summary-result.html")
 
 @app.route("/quiz", methods=["POST", "GET"])
 def quiz():
     return render_template('quiz.html')
 
-@app.route("/quiz-doc", methods=["POST", "GET"])
-def quiz_doc():
+@app.route("/quiz-result", methods=["POST", "GET"])
+def quiz_result():
     if (request.method == "POST"):
-
         file_upload = request.files.get("file")
 
+        if file_upload.mimetype.startswith("audio/") or file_upload.mimetype.startswith("video/"):
+            with tempfile.NamedTemporaryFile(delete=False, suffix=os.path.splitext(file_upload.filename)[1]) as tmp:
+                file_upload.save(tmp.name)
+                tmp_path = tmp.name
+            try:
+                result = get_vid_aud_quiz(tmp_path)
+            finally:
+                os.remove(tmp_path)
+        elif file_upload.mimetype == "application/pdf":
+            result = get_doc_quiz(file_upload)
+
+        template = f"<p class='animated' id='quiz'>{result}</p>"
+        return render_template('quiz-result.html') + template
+    
+    return render_template("quiz-result.html")
+
+@app.route("/extract", methods=["POST", "GET"])
+def extract():
+    return render_template('extract.html')
+
+@app.route("/extract-result", methods=["POST", "GET"])
+def extract_result():
+    if (request.method == "POST"):
+        file_upload = request.files.get("file")
+        result = get_img_txt(file_upload)
+        template = f"<p class='animated' id='extracted'>{result}</p>"
+        return render_template('extract-result.html') + template
+    
+    return render_template("extract-result.html")
+
+@app.route("/transcribe", methods=["POST", "GET"])
+def transcribe():
+    return render_template('transcribe.html')
+
+@app.route("/transcribe-result", methods=["POST", "GET"])
+def transcribe_result():
+    if (request.method == "POST"):
+        file_upload = request.files.get("file")
         with tempfile.NamedTemporaryFile(delete=False, suffix=os.path.splitext(file_upload.filename)[1]) as tmp:
             file_upload.save(tmp.name)
             tmp_path = tmp.name
-        
+
         try:
-            result = get_doc_quiz(tmp_path).replace('\n', '<br>')
-            result = result.replace('**', '')
-
-            template = f"<p class='animated' id='quiz'>{result}</p>"
-
-            return render_template('quiz-doc.html') + template
-    
+            result = get_transcript(tmp_path)
+            template = f"<p class='animated' id='transcription'>{result}</p>"
+            return render_template('transcribe-result.html') + template
         finally:
             os.remove(tmp_path)
 
-    return render_template("quiz-doc.html")
+    return render_template("transcribe-result.html")
 
 if __name__ == "__main__":
     app.run(debug=True)
